Travis CI: Rely on install script by Richard Weickelt
Seems to be rather more convenient and hopefully faster than relying on
a PPA (as done on Linux) or extracting stuff from the huge Qt installer
(as done on macOS).

Also removes the need to compile Qbs ourselves. Now the Qbs released
along with Qt Creator is used.

At the same time the builds are now updated from Qt 5.12.3 to 5.12.4.